Minor changes * Rename 'my_swaps_db::MySwaps' to 'my_swaps_storage::MySwapsStorage' * Rename swap_db.rs to swap_wasm_db.rs
